ci: GitLab pipeline for the cv-builder image.
This is the first of a few commits for creating
the GitLab CI pipeline for building and publishing my CV.
This commit adds a pipeline for building and publishing the cv-builder
docker image to the GitLab container registry.

Additions:

- The cv-builder docker image: This commit adds a Dockerfile for
building the cv-builder docker image. The image is based on the
golang (buster) image and installs ConTeXt (standalone version)
and the Carlito font.

- The gitlab-ci.yml file: This commit adds the gitlab-ci.yml file for
running the pipeline for building and publishing the docker image
to the GitLab Container Registry.

Changes:

- The Makefile now includes targets for building and publishing docker images.