feat: build the magefile into a binary
This commit builds the magefile into a binary during the docker build
stage.
The Dockerfile is updated to include a multi-stage build where the first
stage compiles the binary and the second stage builds the final
cv-builder image.
The final image is now based on debian:buster instead of
golang:<VERSION>-buster which significantly reduces the image's size.